Alignment: fix publisher put encoding fallback
